@@ -34,73 +33,71 @@
     "author": "Ron Penton",
     "license": "MIT",
     "bugs": {
-        "url": "https://github.com/RonPenton/NotaMUD/issues"
+        "url": "https://github.com/RonPenton/generalmud/issues"
     },
-    "homepage": "https://github.com/RonPenton/NotaMUD#readme",
-    "dependencies": {
-        "@types/babel-types": "^7.0.8",
-        "@types/babylon": "^6.16.5",
-        "@types/bcrypt-nodejs": "0.0.31",
-        "@types/body-parser": "^1.19.0",
-        "@types/connect-flash": "0.0.35",
-        "@types/cookie-parser": "^1.4.2",
-        "@types/deep-freeze": "^0.1.2",
-        "@types/express": "^4.17.7",
-        "@types/express-session": "^1.17.0",
-        "@types/lodash": "^4.14.158",
-        "@types/morgan": "^1.9.1",
-        "@types/node": "^14.0.26",
-        "@types/passport": "^1.0.4",
-        "@types/passport-local": "^1.0.33",
-        "@types/ramda": "^0.27.12",
-        "@types/react": "^16.9.43",
-        "@types/react-dom": "^16.9.8",
-        "@types/socket.io": "^2.1.10",
-        "@types/socket.io-client": "^1.4.33",
-        "aws-sdk": "^2.721.0",
-        "babel-types": "^6.26.0",
-        "babylon": "^6.18.0",
-        "bcrypt-nodejs": "0.0.3",
-        "body-parser": "^1.19.0",
-        "connect-dynamodb": "^2.0.4",
-        "connect-flash": "^0.1.1",
-        "cookie-parser": "^1.4.5",
-        "copyfiles": "^2.3.0",
-        "css-loader": "^4.0.0",
-        "decko": "^1.2.0",
-        "deep-freeze": "0.0.1",
-        "dexie": "^3.0.1",
-        "ejs": "^3.1.3",
-        "express": "^4.17.1",
-        "express-session": "^1.17.1",
-        "foundation-sites": "^6.6.3",
-        "lodash": "^4.17.19",
-        "moment": "^2.27.0",
-        "morgan": "^1.10.0",
-        "node-sass": "^4.14.1",
-        "passport": "^0.4.1",
-        "passport-local": "^1.0.0",
-        "passport.socketio": "^3.7.0",
-        "ramda": "^0.27.0",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1",
-        "react-foundation": "^0.9.6",
-        "requirejs": "^2.3.6",
-        "rimraf": "^3.0.2",
-        "sass-loader": "^9.0.2",
-        "socket.io": "^2.3.0",
-        "socket.io-client": "^2.3.0",
-        "style-loader": "^1.2.1",
-        "ts-keycode-enum": "^1.0.6",
-        "ts-loader": "^8.0.1",
-        "ts-priority-queue": "^0.1.1",
-        "tslint": "^6.1.2",
-        "typescript": "^3.9.7",
-        "webpack": "^4.44.0"
-    },
+    "homepage": "https://github.com/RonPenton/generalmud#readme",
     "devDependencies": {
-        "mini-css-extract-plugin": "^0.9.0",
-        "tsconfig-paths-webpack-plugin": "^3.2.0",
-        "webpack-cli": "^3.3.12"
+        "@types/bcryptjs": "^2.4.2",
+        "@types/clone": "^2.1.1",
+        "@types/connect-flash": "^0.0.37",
+        "@types/connect-pg-simple": "^7.0.0",
+        "@types/cookie-parser": "^1.4.3",
+        "@types/csvtojson": "^2.0.0",
+        "@types/decimal.js": "^7.4.0",
+        "@types/deep-freeze": "^0.1.2",
+        "@types/express": "^4.17.14",
+        "@types/express-session": "^1.17.5",
+        "@types/morgan": "^1.9.3",
+        "@types/passport": "^1.0.11",
+        "@types/passport-local": "^1.0.34",
+        "@types/react": "^18.0.21",
+        "@types/react-dom": "^18.0.6",
+        "@types/socket.io": "^3.0.2",
+        "copyfiles": "^2.4.1",
+        "cross-env": "^7.0.3",
+        "css-loader": "^6.7.1",
+        "mini-css-extract-plugin": "^2.6.1",
+        "node-adodb": "^5.0.3",
+        "nodemon": "^2.0.20",
+        "rimraf": "^3.0.2",
+        "sass": "^1.55.0",
+        "sass-loader": "^13.1.0",
+        "ts-loader": "^9.4.1",
+        "ts-node": "^10.9.1",
+        "tsconfig-paths-webpack-plugin": "^4.0.0",
+        "typescript": "^4.8.4",
+        "webpack": "^5.74.0",
+        "webpack-cli": "^4.10.0",
+        "webpack-dev-middleware": "^5.3.3",
+        "webpack-hot-middleware": "^2.25.2"
+    },
+    "dependencies": {
+        "bcryptjs": "^2.4.3",
+        "big-number": "^2.0.0",
+        "clone": "^2.1.2",
+        "connect-flash": "^0.1.1",
+        "connect-pg-simple": "^7.0.0",
+        "cookie-parser": "^1.4.6",
+        "csvtojson": "^2.0.10",
+        "date-fns": "^2.29.3",
+        "decimal.js": "^10.4.2",
+        "decko": "^1.2.0",
+        "deep-freeze": "^0.0.1",
+        "dotenv-json2": "^1.1.0",
+        "ejs": "^3.1.8",
+        "express": "^4.18.2",
+        "express-session": "^1.17.3",
+        "foundation-sites": "^6.7.5",
+        "morgan": "^1.10.0",
+        "passport": "^0.6.0",
+        "passport-local": "^1.0.0",
+        "pg-promise": "^10.12.0",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "socket.io": "^4.5.2",
+        "socket.io-client": "^4.5.3",
+        "ts-dotenv-parser": "^1.1.0",
+        "ts-priority-queue": "^0.1.1",
+        "tsc-utils": "^2.3.0"
     }
 }
