@@ -1,15 +1,11 @@
-import { GameContext } from '../App';
 import React from 'react';
-import { create, ShowTimestamp, TimeStamp } from './index';
-import * as Messages from '../../server/messages';
+import { ChildrenProps } from '../foundation/top-bar';
+import { createClientCommand } from './base';
 
-export const command = create('error',
-    (message: Messages.TimeStamped<Messages.Error>, context: GameContext) => {
-        context.addOutput(<ErrorComponent time={message.timeStampStr}>{message.message}</ErrorComponent>)
-    });
+createClientCommand('error', (message, context) => {
+    context.addOutput(<ErrorComponent>{message.message.text}</ErrorComponent>)
+});
 
-
-
-export const ErrorComponent: React.SFC<ShowTimestamp> = (props) => {
-    return <div className="message error"><TimeStamp time={props.time} />{props.children}</div>
+export const ErrorComponent: React.FC<ChildrenProps> = (props) => {
+    return <div className="message error">{props.children}</div>
 }
