@@ -1,32 +1,21 @@
 import React from 'react';
 import { ErrorComponent } from './error';
 import { GameContext } from '../App';
-import { Message, TimeStamped } from '../../server/messages';
-import moment from 'moment';
+import { ChildrenProps } from '../foundation/top-bar';
+import { MessageName, MessagePacket } from '../../server/messages';
+import { getClientCommand } from './base';
 
-export type ExecuteFunction<T extends Message> = (message: T, context: GameContext) => (Promise<boolean> | void);
-export type ExecuteFunctionPromise<T extends Message> = (message: T, context: GameContext) => Promise<boolean>;
+import './error';
+import './connected';
+import './disconnected';
+import './system';
+import './talk-global';
+import './talk-room';
+import './room-description';
+import './actor-moved';
 
-export type Command<T extends Message = Message> = {
-    name: string,
-    execute: ExecuteFunction<T>
-}
-
-export type Commands = Map<string, Command>;
-const commands = new Map<string, Command>();
-export function install(command: Command) {
-    commands.set(command.name, command);
-}
-
-export type InstallFunction = (command: Command) => void;
-export type CreateFunction = (install: InstallFunction) => void;
-
-export function create<T extends Message>(name: string, execute: ExecuteFunction<T>): Command<T> {
-    return { name, execute };
-}
-
-export function handle(message: TimeStamped<Message>, context: GameContext) {
-    const command = commands.get(message.type);
+export function handle<T extends MessageName>(message: MessagePacket<T>, context: GameContext) {
+    const command = getClientCommand(message.type);
     if (!command) {
         const error = `Message not supported: ${message.type}`;
         context.addOutput(<ErrorComponent>{error}</ErrorComponent>)
@@ -36,14 +25,6 @@ export function handle(message: TimeStamped<Message>, context: GameContext) {
     command.execute(message, context);
 }
 
-export type ShowTimestamp = { time?: string };
-export const TimeStamp: React.SFC<ShowTimestamp> = props => {
-    if (!props.time)
-        return null;
-    const time = moment(props.time);
-    return <span className="timestamp">{`[${time.format('LT')}] `}</span>
-}
-
-export const Generic: React.SFC<ShowTimestamp> = (props) => {
-    return <div className="message generic"><TimeStamp time={props.time} />{props.children}</div>
+export const Generic: React.FC<ChildrenProps> = (props) => {
+    return <div className="message generic">{props.children}</div>
 }
