@@ -1,29 +1,34 @@
 import React from 'react';
 
-export const TopBar: React.SFC<{}> = (props) => {
+export interface ChildrenProps {
+    children?: React.ReactNode
+}
+
+export const TopBar: React.FC<ChildrenProps> = (props) => {
     return <div className="top-bar">{props.children}</div>;
 }
 
-export const TopBarLeft: React.SFC<{}> = (props) => {
+export const TopBarLeft: React.FC<ChildrenProps> = (props) => {
     return <div className="top-bar-left">{props.children}</div>;
 }
 
-export const TopBarRight: React.SFC<{}> = (props) => {
+export const TopBarRight: React.FC<ChildrenProps> = (props) => {
     return <div className="top-bar-right">{props.children}</div>;
 }
 
-export const TopBarMenu: React.SFC<{}> = (props) => {
+export const TopBarMenu: React.FC<ChildrenProps> = (props) => {
     return <ul className="menu">{props.children}</ul>;
 }
 
-export const TopBarMenuTitle: React.SFC<{}> = (props) => {
+export const TopBarMenuTitle: React.FC<ChildrenProps> = (props) => {
     return <li className="menu-text">{props.children}</li>;
 }
 
 export interface TopBarMenuItemProps {
     href: string;
+    children?: React.ReactNode
 }
 
-export const TopBarMenuItem: React.SFC<TopBarMenuItemProps> = (props) => {
+export const TopBarMenuItem: React.FC<TopBarMenuItemProps> = (props) => {
     return <li><a href={props.href}>{props.children}</a></li>;
 }
