@@ -1,12 +1,24 @@
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
+const webpack = require('webpack');
+
+const devMode = process.env.NODE_ENV !== 'production';
+
+const mode = devMode ? 'development' : 'production';
 
 var extractPlugin = new MiniCssExtractPlugin({
     filename: 'bundle.css'
 });
 
+let main = ['./src/client/App.tsx'];
+if (devMode) {
+    main.unshift('webpack-hot-middleware/client?reload=true');
+}
+
+
 module.exports = {
-    entry: "./src/client/App.tsx",
+    entry: { main },
+
     output: {
         filename: "client.js",
         path: __dirname + "/../../build/public"
@@ -37,6 +49,7 @@ module.exports = {
 
     plugins: [
         extractPlugin,
+        new webpack.HotModuleReplacementPlugin()
     ],
 
 };
